+++ pkg/deploy/registry/pluginregistry_test.go
@@ -12,116 +12,37 @@
 package registry
 
 import (
-	"os"
-
-	"github.com/eclipse-che/che-operator/pkg/util"
+	"context"
 
 	"github.com/eclipse-che/che-operator/pkg/deploy"
+	"github.com/eclipse-che/che-operator/pkg/util"
 
 	orgv1 "github.com/eclipse-che/che-operator/pkg/apis/org/v1"
+	routev1 "github.com/openshift/api/route/v1"
+	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/kubernetes/scheme"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
-	"sigs.k8s.io/controller-runtime/pkg/log/zap"
-	logf "sigs.k8s.io/controller-runtime/pkg/runtime/log"
 
 	"testing"
 )
 
-func TestPluginRegistryDeployment(t *testing.T) {
-	type testCase struct {
-		name          string
-		initObjects   []runtime.Object
-		memoryLimit   string
-		memoryRequest string
-		cpuRequest    string
-		cpuLimit      string
-		cheCluster    *orgv1.CheCluster
-	}
-
-	testCases := []testCase{
-		{
-			name:          "Test default limits",
-			initObjects:   []runtime.Object{},
-			memoryLimit:   deploy.DefaultPluginRegistryMemoryLimit,
-			memoryRequest: deploy.DefaultPluginRegistryMemoryRequest,
-			cpuLimit:      deploy.DefaultPluginRegistryCpuLimit,
-			cpuRequest:    deploy.DefaultPluginRegistryCpuRequest,
-			cheCluster: &orgv1.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-				},
-			},
-		},
-		{
-			name:          "Test custom limits",
-			initObjects:   []runtime.Object{},
-			cpuLimit:      "250m",
-			cpuRequest:    "150m",
-			memoryLimit:   "250Mi",
-			memoryRequest: "150Mi",
-			cheCluster: &orgv1.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-				},
-				Spec: orgv1.CheClusterSpec{
-					Server: orgv1.CheClusterSpecServer{
-						PluginRegistryCpuLimit:      "250m",
-						PluginRegistryCpuRequest:    "150m",
-						PluginRegistryMemoryLimit:   "250Mi",
-						PluginRegistryMemoryRequest: "150Mi",
-					},
-				},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			logf.SetLogger(zap.LoggerTo(os.Stdout, true))
-			orgv1.SchemeBuilder.AddToScheme(scheme.Scheme)
-			testCase.initObjects = append(testCase.initObjects)
-			cli := fake.NewFakeClientWithScheme(scheme.Scheme, testCase.initObjects...)
-
-			deployContext := &deploy.DeployContext{
-				CheCluster: testCase.cheCluster,
-				ClusterAPI: deploy.ClusterAPI{
-					Client: cli,
-					Scheme: scheme.Scheme,
-				},
-				Proxy: &deploy.Proxy{},
-			}
-
-			pluginregistry := NewPluginRegistry(deployContext)
-			deployment := pluginregistry.GetPluginRegistryDeploymentSpec()
-
-			util.CompareResources(deployment,
-				util.TestExpectedResources{
-					MemoryLimit:   testCase.memoryLimit,
-					MemoryRequest: testCase.memoryRequest,
-					CpuRequest:    testCase.cpuRequest,
-					CpuLimit:      testCase.cpuLimit,
-				},
-				t)
-
-			util.ValidateSecurityContext(deployment, t)
-		})
-	}
-}
-
 func TestPluginRegistrySyncAll(t *testing.T) {
+	cheCluster := &orgv1.CheCluster{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: "eclipse-che",
+			Name:      "eclipse-che",
+		},
+	}
+
 	orgv1.SchemeBuilder.AddToScheme(scheme.Scheme)
 	corev1.SchemeBuilder.AddToScheme(scheme.Scheme)
-	cli := fake.NewFakeClientWithScheme(scheme.Scheme)
+	routev1.AddToScheme(scheme.Scheme)
+	cli := fake.NewFakeClientWithScheme(scheme.Scheme, cheCluster)
 	deployContext := &deploy.DeployContext{
-		CheCluster: &orgv1.CheCluster{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: "eclipse-che",
-				Name:      "eclipse-che",
-			},
-		},
+		CheCluster: cheCluster,
 		ClusterAPI: deploy.ClusterAPI{
 			Client:          cli,
 			NonCachedClient: cli,
@@ -129,9 +50,36 @@ func TestPluginRegistrySyncAll(t *testing.T) {
 		},
 	}
 
+	util.IsOpenShift = true
+
 	pluginregistry := NewPluginRegistry(deployContext)
 	done, err := pluginregistry.SyncAll()
 	if !done || err != nil {
-		t.Fatalf("Failed to sync Devfile Registry: %v", err)
+		t.Fatalf("Failed to sync Plugin Registry: %v", err)
+	}
+
+	// check service
+	service := &corev1.Service{}
+	err = cli.Get(context.TODO(), types.NamespacedName{Name: "plugin-registry", Namespace: "eclipse-che"}, service)
+	if err != nil {
+		t.Fatalf("Service not found: %v", err)
+	}
+
+	// check endpoint
+	route := &routev1.Route{}
+	err = cli.Get(context.TODO(), types.NamespacedName{Name: "plugin-registry", Namespace: "eclipse-che"}, route)
+	if err != nil {
+		t.Fatalf("Route not found: %v", err)
+	}
+
+	// check deployment
+	deployment := &appsv1.Deployment{}
+	err = cli.Get(context.TODO(), types.NamespacedName{Name: "plugin-registry", Namespace: "eclipse-che"}, deployment)
+	if err != nil {
+		t.Fatalf("Deployment not found: %v", err)
+	}
+
+	if cheCluster.Status.PluginRegistryURL == "" {
+		t.Fatalf("Status wasn't updated")
 	}
 }
